+++ src/main/steps/respondent/yourdocuments/alldocuments/police_disclosures/content.ts
@@ -1,6 +1,6 @@
 //import config from 'config';
 //import { getSystemUser } from 'app/auth/user/oidc';
-import { APPLICANT_ORDERS_FROM_THE_COURT } from '../../../../../../main/steps/urls';
+import { POLICE_DISCLOSURE } from '../../../../../../main/steps/urls';
 import { TranslationFn } from '../../../../../app/controller/GetController';
 import { FormContent } from '../../../../../app/form/Form';
 
@@ -49,18 +49,21 @@ export const form: FormContent = {
 
 export const generateContent: TranslationFn = content => {
   const translations = languages[content.language]();
-  console.log('hello' + content.name);
-
   const orders: object[] = [];
-  for (const doc of content.userCase?.orderCollection || []) {
-    const uid = doc.value.orderDocument.document_url.substring(
-      doc.value.orderDocument.document_url.lastIndexOf('/') + 1
-    );
-    orders.push({
-      href: `${APPLICANT_ORDERS_FROM_THE_COURT}/${uid}`,
-      createdDate: doc.value.otherDetails.orderCreatedDate,
-      fileName: doc.value.orderDocument.document_filename,
-    });
+  for (const doc of content.userCase?.citizenUploadedDocumentList || []) {
+    if (
+      doc.value.documentType === 'Police reports' &&
+      content.byApplicant === doc.value.isApplicant
+    ) {
+      const uid = doc.value.citizenDocument.document_url.substring(
+        doc.value.citizenDocument.document_url.lastIndexOf('/') + 1
+      );
+      orders.push({
+        href: `${POLICE_DISCLOSURE}/${uid}`,
+        createdDate: doc.value.documentUploadedDate,
+        fileName: doc.value.citizenDocument.document_filename,
+      });
+    }
   }
 
   return {
