--- src/main/steps/respondent/yourdocuments/alldocuments/policereports/content.test.ts
@@ -1,54 +1,0 @@
-import languageAssertions from '../../../../../../test/unit/utils/languageAssertions';
-//import { FormContent /*, FormFields,  FormOptions*/ } from '../../../../../app/form/Form';
-import { CommonContent } from '../../../../common/common.content';
-
-import { generateContent } from './content';
-
-const enContent = {
-  section: 'All documents',
-  title: 'Letters from school',
-  threeHint: 'This is a 8 character code',
-  summaryText: 'Contacts for help',
-  caseNumber: 'Case number',
-  continue: 'Go back',
-};
-
-const cyContent = {
-  section: 'All documents',
-  title: 'Letters from school',
-  threeHint: 'This is a 8 character code',
-  summaryText: 'Contacts for help',
-  caseNumber: 'Case number',
-  continue: 'Go back',
-};
-
-jest.mock('../../../../../app/form/validation');
-/* eslint-disable @typescript-eslint/ban-types */
-describe('citizen-home content', () => {
-  const commonContent = { language: 'en' } as CommonContent;
-  let generatedContent;
-  //let form;
-  //let fields;
-  beforeEach(() => {
-    generatedContent = generateContent(commonContent);
-    //form = generatedContent.form as FormContent;
-    //fields = form.fields as FormFields;
-  });
-
-  test('should return correct english content', () => {
-    expect(generatedContent.title).toEqual('Letters from school');
-    expect(generatedContent.section).toEqual('All documents');
-    expect(generatedContent.summaryText).toEqual('Contacts for help');
-  });
-
-  // eslint-disable-next-line jest/expect-expect
-  test('should return correct english content Data', () => {
-    languageAssertions('en', enContent, () => generateContent(commonContent));
-  });
-
-  // eslint-disable-next-line jest/expect-expect
-  test('should return correct welsh content', () => {
-    languageAssertions('cy', cyContent, () => generateContent({ ...commonContent, language: 'cy' }));
-  });
-});
-/* eslint-enable @typescript-eslint/ban-types */
