--- src/main/steps/respondent/yourdocuments/alldocuments/policereports/content.ts
@@ -1,73 +1,0 @@
-//import config from 'config';
-//import { getSystemUser } from 'app/auth/user/oidc';
-import { POLICE_REPORTS } from '../../../../urls';
-import { TranslationFn } from '../../../../../app/controller/GetController';
-import { FormContent } from '../../../../../app/form/Form';
-
-const en = () => {
-  return {
-    section: 'All documents',
-    title: 'Letters from school',
-    threeHint: 'This is a 8 character code',
-    summaryText: 'Contacts for help',
-    caseNumber: 'Case number',
-    continue: 'Go back',
-  };
-};
-
-const cy: typeof en = () => {
-  return {
-    section: 'All documents',
-    title: 'Letters from school',
-    threeHint: 'This is a 8 character code',
-    summaryText: 'Contacts for help',
-    caseNumber: 'Case number',
-    continue: 'Go back',
-  };
-};
-
-const languages = {
-  en,
-  cy,
-};
-
-export const form: FormContent = {
-  fields: userCase => {
-    return {
-      caseNumber: {
-        label: l => l.caseNumber + '' + userCase.caseCode,
-        type: 'hidden',
-        labelHidden: true,
-      },
-    };
-  },
-  submit: {
-    text: l => l.continue,
-    classes: 'govuk-button--secondary',
-  },
-};
-
-export const generateContent: TranslationFn = content => {
-  const translations = languages[content.language]();
-  const orders: object[] = [];
-  for (const doc of content.userCase?.citizenUploadedDocumentList || []) {
-    if (
-      doc.value.documentType === 'Emails, screenshots, images and other media files' &&
-      content.byApplicant === doc.value.isApplicant
-    ) {
-      const uid = doc.value.citizenDocument.document_url.substring(
-        doc.value.citizenDocument.document_url.lastIndexOf('/') + 1
-      );
-      orders.push({
-        href: `${POLICE_REPORTS}/${uid}`,
-        createdDate: doc.value.documentUploadedDate,
-        fileName: doc.value.citizenDocument.document_filename,
-      });
-    }
-  }
-
-  return {
-    ...translations,
-    orders,
-  };
-};
